requirements.txt
numpy==1.19.2
Pillow==9.2.0
matplotlib==3.3.2
